requirements.txt
python-telegram-bot>=21.0
supabase>=2.10.0
openai>=1.12.0
apscheduler>=3.10.4
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pytz>=2024.1
